+++ CHANGELOG.md
@@ -2,3 +2,4 @@ v1.0.0
 - * Breaking Changes *
 - Added Ephemeris class and implemented ES6 modules throughout codebase
 - Requres month range 0 - 11 on instantiation
+- Fixes Moon Phase `phaseQuarter` bug, and adds `phaseQuarterString` property to the moon result with a descriptor of the phase.